refactor: some more cleanup
