--- mcp-config-example.json
@@ -1,12 +1,0 @@
-{
-    "mcpServers": {
-        "leetcode-extended": {
-            "command": "node",
-            "args": [
-                "/Users/svetlana/Manual/leetcode/leetcode-mcp-extended/build/index.js",
-                "--site",
-                "global"
-            ]
-        }
-    }
-}
