+++ package.json
@@ -46,7 +46,7 @@
         ]
     },
     "dependencies": {
-        "@modelcontextprotocol/sdk": "^1.8.0",
+        "@modelcontextprotocol/sdk": "^1.25.2",
         "axios": "^1.13.2",
         "leetcode-query": "^2.0.0",
         "minimist": "^1.2.8",
