--- smithery.yaml
@@ -1,31 +1,0 @@
-# Smithery configuration file: https://smithery.ai/docs/config#smitheryyaml
-
-startCommand:
-    type: stdio
-    configSchema:
-        # JSON Schema defining the configuration options for the MCP.
-        type: object
-        properties:
-            site:
-                type: string
-                default: global
-                description: "LeetCode site to target: 'global' for leetcode.com, or 'cn' for leetcode.cn"
-            session:
-                type: string
-                description: Optional session cookie to authenticate with LeetCode API
-    commandFunction:
-        # A JS function that produces the CLI command based on the given config to start the MCP on stdio.
-        |-
-        (config) => {
-            const args = ['build/index.js', '--site', config.site];
-            if (config.session && config.session.trim() !== '') {
-                args.push('--session', config.session);
-            }
-            return {
-                command: 'node',
-                args
-            };
-        }
-    exampleConfig:
-        site: global
-        session: leetcode_session_cookie_value
